manager/requirements.txt
aioredis==2.0.1
alabaster==0.7.12
asgiref==3.8.1
asn1crypto==1.3.0
astropy[recommended]==7.1.0
async-timeout==3.0.1
atomicwrites==1.4.1
attrs==25.3.0
autobahn==23.6.2
Automat==20.2.0
Babel==2.9.1
certifi==2024.7.4
cffi==2.0.0
channels==4.1.0
channels-redis==4.2.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.1
daphne==4.2.1
Django==5.0.13
django-auth-ldap==4.8.0
django-cors-headers==4.4.0
django-webpack-loader==0.7.0
djangorestframework==3.15.2
docutils==0.16
drf-yasg==1.21.7
entrypoints==0.3
freezegun==1.5.5
hiredis==3.0.0
hyperlink==21.0.0
idna==3.7
imagesize==1.2.0
importlib-metadata==1.5.0
incremental==24.7.2
inflection==0.3.1
itypes==1.1.0
Jinja2==3.1.6
jsonschema==4.25.1
MarkupSafe==2.1.5
mccabe==0.6.1
mistune<2.0.0
more-itertools==8.2.0
msgpack==1.0.0
numpy==2.3.3
packaging==24.1
pip-licenses==2.1.1
Pillow==11.3.0
pluggy==1.6.0
psycopg==3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.15.0
PyHamcrest==2.0.2
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-django==4.11.1
pytest-env==1.1.5
python-dateutil==2.9.0.post0
python-ldap==3.4.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.4
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
six==1.17.0
snowballstemmer==2.0.0
sqlparse==0.5.0
standard-imghdr==3.13.0
Twisted==24.7.0
txaio==23.1.1
uritemplate==3.0.1
urllib3==2.5.0
uvicorn[standard]==0.24.0.post1
wcwidth==0.1.8
zipp==3.19.1
zope.interface==6.4
